Matched 11 functions in handcart and switched to bitfields for a few variables in Actor

+++ include/gc/dialogcamera.h
@@ -2,6 +2,6 @@
 #define __GC_DIALOG_CAMERA_H__
 #include "common.h"
 
-void _gcdialogcamera_entrypoint_3(void*, s32);
+void _gcdialogcamera_entrypoint_3(void*, f32);
 
 #endif
